condenses read files
