+++ py4DSTEM/io/__init__.py
@@ -1,7 +1,6 @@
 from py4DSTEM.io import datastructure
 from py4DSTEM.io.datastructure import *
 from py4DSTEM.io.read import *
-from py4DSTEM.io.import_file import *
 from py4DSTEM.io.native.read import print_h5_tree
 from py4DSTEM.io.native.write import save
 from py4DSTEM.io.google_drive_downloader import download_file_from_google_drive, get_sample_data_ids
